+++ changelog.md
@@ -1,7 +1,10 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
-## V1.14.29 - 2020-09-11
+## V1.14.31 - 2020-09-17
+- Added ESSID to WiFi platform information
+
+## V1.14.30 - 2020-09-11
 - Updagte to MQTT support to allow list of strings to be a parameter
 - Added advanced parameter to better support weak wifi signals on serial over TCP
 - Made Evolution 2.0 Ambient temp sensor display respond to Use Metric setting